feat: full platform update — contracts, marketplace, battle UI, agents, API routes
Contracts: TeamBattleEngine, FrostbiteMarketplace, GameEngine, BattleEngineProxy,
AgentRegistry updates, RewardVault, Tournament enhancements.

Frontend: marketplace pages, battle page overhaul, agent arena/dashboard,
leaderboard, chat, homepage, pixel theme, particle background, docs/privacy/terms.

API: v1 routes for battles, feed, follow, heartbeat, leaderboard, balance,
warriors, tournaments, agent marketplace, reputation, challenge, rivals.

Scripts: deploy, verify, configure, test-mint, upgrade-audit-fixes.

+++ types/ethers-contracts/factories/BattleEngine.sol/IArenaWarrior__factory.ts
@@ -44,9 +44,9 @@
             "type": "uint8"
           },
           {
-            "internalType": "uint8",
+            "internalType": "uint16",
             "name": "level",
-            "type": "uint8"
+            "type": "uint16"
           },
           {
             "internalType": "uint256",
